split tests into separate files to test jenkins build
